bpf: Teach verifier about non-size constant arguments
Currently, the verifier has support for various arguments that either
describe the size of the memory being passed in to a helper, or describe
the size of the memory being returned. When a constant is passed in like
this, it is assumed for the purposes of precision tracking that if the
value in the already explored safe state is within the value in current
state, it would fine to prune the search.

While this holds well for size arguments, arguments where each value may
denote a distinct meaning and needs to be verified separately needs more
work. Search can only be pruned if both are constant values and both are
equal. In all other cases, it would be incorrect to treat those two
precise registers as equivalent if the new value satisfies the old one
(i.e. old <= cur).

Hence, make the register precision marker tri-state. There are now three
values that reg->precise takes: NOT_PRECISE, PRECISE, PRECISE_ABSOLUTE.

Both PRECISE and PRECISE_ABSOLUTE are 'true' values. PRECISE_ABSOLUTE
affects how regsafe decides whether both registers are equivalent for
the purposes of verifier state equivalence. When it sees that one
register has reg->precise == PRECISE_ABSOLUTE, unless both are absolute,
it will return false. When both are, it returns true only when both are
const and both have the same value. Otherwise, for PRECISE case it falls
back to the default check that is present now (i.e. thinking that we're
talking about sizes).

This is required as a future patch introduces a BPF memory allocator
interface, where we take the program BTF's type ID as an argument. Each
distinct type ID may result in the returned pointer obtaining a
different size, hence precision tracking is needed, and pruning cannot
just happen when the old value is within the current value. It must only
happen when the type ID is equal. The type ID will always correspond to
prog->aux->btf hence actual type match is not required.

Finally, change mark_chain_precision to mark_chain_precision_absolute
for kfuncs constant non-size scalar arguments (tagged with __k suffix).

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>