bpf: Eschew extra delete_work work_struct for bpf_work
We recently added a delete_work work_struct to bpf_async_cb structure.
This is meant to schedule deletion related work for the async object.
It shares storage with rcu_head, and is used for scheduling work timer
deletion. The same struct can also be used for workqueue deletion,
therefore remove the extra delete_work work_struct in bpf_work and unify
usage.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>