qspinlock: Move common helpers to header
Move qspinlock functions that encode, decode, and manipulate the tail
word, manipulate the pending and locked bits, and other trivial defines
that will be useful for implementing a modified qspinlock slow path. We
want to share the layout of the qspinlock word, and its representation.

However, we will not share the exact qnode type, hence make the existing
code independent of the concrete type. It can be set using _Q_NODE_TYPE
macro when including the header, to allow changing the underlying type.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>