rqspinlock: Add support for timeouts
Introduce policy macro RES_CHECK_TIMEOUT which can be used to detect
when the timeout has expired for the slow path to return an error. It
depends on being passed two variables initialized to 0: ts, ret. The
'ts' parameter is of type rqspinlock_timeout.

The 'spin' member is used to amortize the cost of checking time by
dispatching to the implementation every 64k iterations. The 'end' member
is used to keep track of the timestamp that denotes the end of the
waiting period. The 'ret' parameter denotes the status of the timeout,
and can be checked in the slow path to detect timeouts after waiting
loops.

The 'timeout' member is used to store the timeout for each waiting loop,
that is passed down from the caller of the slow path function.  Use the
RES_INIT_TIMEOUT macro to initialize it.

This macro will be used as a condition for waiting loops in the slow
path.  Since each waiting loop applies a fresh timeout using the same
rqspinlock_timeout, we add a new RES_RESET_TIMEOUT as well to ensure the
values can be easily reinitialized to the default state.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>