bpf, RSL: Add infra to support timeouts
Introduce policy macro RES_CHECK_TIMEOUT which can be used to detect
when the timeout has expired for the slow path to return an error. It
depends on being passed three variables initialized to 0: spin, end,
ret. The 'spin' variable is used to amortize the cost of checking time
by dispatching to the implementation every 16k iterations. The 'end'
variable is used to keep track of the timestamp that denotes the end of
the waiting period. The 'ret' variable denotes the status of the
timeout, and can be checked in the slow path to detect timeouts after
waiting loops. We also add a lock event 'rsl_lock_timeout' which will be
used by subsequent patches to capture timeout events.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>