locking: Return value in arch_mcs_spin_lock_contended
We will need to inspect the return value once the conditional waiting
loop of arch_mcs_spin_lock_contended terminates. To enable this, change
the macro to resolve to a direct smp_cond_load_acquire expression, so
that the return value can be inspected if necessary. This will be used
to disambiguate the MCS node's locked value in later patches.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>