bpf, RSL: Protect waiters in queue from stalls
Implement the wait queue cleanup algorithm for RSL. There are three
forms of waiters in the original queued spin lock algorithm. The first
is the waiter which acquires the pending bit and spins on the lock word
without forming a wait queue. The second is the head waiter that is the
first waiter heading the wait queue. The third form is of all the
non-head waiters queued behind the head, waiting to be signalled to
overtake the responsibility of the head (i.e. spin for pending and
locked bits to go to 0).

In this commit, we are concerned with the second and third kind. First,
we augment the waiting loop of the head of the wait queue with a
timeout. When this timeout happens, all waiters part of the wait queue
will abort their lock acquisition attempts. This happens in three steps.
First, the head breaks out of its loop waiting for pending and locked
bits to turn to 0, and non-head waiters break out of their MCS node spin
(more on that later). Next, every waiter (head or non-head) attempts to
check whether they are also the tail waiter, in such a case they attempt
to zero out the tail word and allow a new queue to be built up for this
lock. If they succeed, they have no one to signal next in the queue to
stop spinning. Otherwise, they signal the MCS node of the next waiter to
break out of its spin and try resetting the tail word back to 0. This
goes on until the tail waiter is found. In case of races, the new tail
will be responsible for performing the same task, as the old tail will
then fail to reset the tail word and wait for its next pointer to be
updated before it signals the new tail to do the same.

TODO: Figure

Lastly, all of these waiters release the rqnode and return to the
caller. This patch underscores the point that RSL's timeout does not
apply to each waiter individually, and cannot be relied upon as the
upper bound. It is possible for the RSL waiters to return early from a
failed lock acquisition attempt as soon as stalls are detected.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>