bpf, RSL: Add deadlock detection and recovery
While the timeout logic provides guarantees for the waiter's forward
progress, the time until a stalling waiter unblocks can still be long.
The default timeout of 32 ms can be excessively long for some use cases.
Additionally, custom timeouts may exacerbate recovery time.

Introduce logic to RSL to detect common cases of deadlocks and perform
quicker recovery. This is done by dividing the time from entry into
the locking slow path until the timeout into intervals of 1 ms. Then,
after each interval elapses, deadlock detection is performed, while also
polling the lock word to ensure we can quickly break out of the the
detection logic and proceed with lock acquisition.

AA deadlocks are only checked in the first interval, as their result
will remain unchanged in subsequent intervals. ABBA deadlocks are
searched for continuously in each interval, as we may miss detection due
to races in intermediate intervals.

A table is maintained per-CPU where the entry at the bottom denotes a
lock being waited for or already taken. Entries coming before it until
the top denote locks that are already held. The current CPU's table can
thus be looked at to detect AA deadlocks. The tables from other CPUs can
be looked at to discover ABBA situations. Writer presence when reading
through tables is detected using sequence locks. Since in a deadlock
detection scenario the table will remain relatively stable, retries are
not attempted and the detection for that CPU is deferred to the
subsequent interval. Finally, when a matching entry for the lock being
taken on the current CPU is found on some other CPU, a deadlock
situation is detected. This function can take a long time, therefore the
lock word is constantly polled in each loop iteration to ensure we can
preempt detection and proceed with lock acquisition.

// Extending lock word size by 4 bytes to record owner CPU can allow
faster detection. It is typically the owner which participates in a ABBA
situation.

// TODO(kkd): Since we only search for held locks in AA and ABBA cases,
we can delay write in dde until exit from slow path.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>