bpf: Refactor active lock management
When bpf_spin_lock was introduced originally, there was deliberation on
whether to use an array of lock IDs, but since bpf_spin_lock is limited
to holding a single lock at any given time, we've been using a single ID
to identify the held lock.

In preparation for introducing resilient spin locks that can be taken
multiple times, introduce support for acquiring multiple lock IDs. For
this purpose, reuse the acquired_refs array and store both lock and
pointer references. We tag the entry with REF_TYPE_PTR or
REF_TYPE_BPF_LOCK to disambiguate and find the relevant entry. The ptr
field is used to track the map_ptr or btf (for bpf_obj_new allocations)
to ensure locks can be matched with protected fields within the same
"allocation", i.e.  bpf_obj_new object or map value.

The struct active_lock is changed to a boolean as the state is part of
the acquired_refs array, and we only need active_lock as a cheap way
of detecting lock presence.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>