bpf: Introduce verifier support for non-arena rqspinlock
Teach the verifier about the conditional lock semantics of
bpf_res_spin_lock, and open up restrictions around holding multiple
locks at a time.

The semantics are that the result of the lock function creates a
conditional lock entry in verifier state, which is in a dual state of
held or not held until the result of bpf_res_spin_lock kfunc is checked.
Thus, operations which require lock protection won't proceed as the lock
may well not be held, and the program won't be able to exit either as
without checking the result, we may have lingering held locks. Once the
result has been checked, the path where the value is non-NULL (i.e. the
lock was acquired successfully), the conditional lock entry in verifier
state is "upgraded" to a proper lock entry, allowing operations lock
protection to succeed. In the path where return value is NULL, the
conditional lock entry is deleted.

To enable this logic, the a new register type LOCK_CONDITION which acts
a non-dereferenceable pointer is introduced. mark_ptr_or_null_regs logic
is special cased for such LOCK_CONDITION values (as it is combined with
PTR_MAYBE_NULL). Whenever bpf_res_spin_lock is called, a conditional
lock entry will be created, and cur->active_lock will be incremented.
In the NULL path, cur->active_lock will be decremented, while in the
non-NULL case, the conditional lock entry in acquired refs will be
upgraded to a proper lock entry (REF_TYPE_RES_LOCK_COND patched to
REF_TYPE_RES_LOCK).

The find_lock_state gains a 'cond' boolean parameter, since when
acquiring locks and finding existing ones, we need to search for
conditional entries to detect possible AA situations, but the same won't
be true when using find_lock_state in check_reg_allocation_locked (i.e.
finding held lock for a given allocation).

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>