rqspinlock: Protect pending->locked hand-over waiters from stalls
When hosted on BPF arena, the lock word can be overwritten, leading us
to observe the 0,1,0 (i.e. pending bit set) state. Typically, this state
is observed when a pending waiter is queued, and the lock owner has
unlocked the lock, which means the pending waiter will use the function
clear_pending_set_locked to transition to acquire the lock. However, due
to corruption, there may be no active pending waiter, and we may keep
spinning for a long time, as _Q_PENDING_LOOPS can be a big value.

Therefore, instrument the waiting loop for the arena slow path to time
the waiting loop and insert a condition to return an error when it
expires. These checks only apply to the arena slow path, so we build on
top of the separate slow path generation logic in previous commit and
abstract checks using macros.

Note that deadlock detection won't apply to this waiting loop, as we
don't grab a held lock entry this early. Therefore, refactor
RES_CHECK_TIMEOUT to allow users to opt out of check_deadlock logic.

We also introduce a new lockevent (rqspinlock_lock_corrupt_timeout)
to distinguish timeout events occuring due to corruption from the
typical ones due to stalled owners or deadlocks.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>