rqspinlock: Sanitize node accesses for arena locks
Due to arena corruption, the tail can randomly point to any node on any
CPU. As such, the current method used to link a waiter's node into the
wait queue (i.e. observing previous node from the xchg_tail and writing
to prev->next) is bound to introduce issues. First, the tail value may
be garbage, hence we need to sanity check the decoded cpu and index from
a given tail. Second, we need to ensure that the previous node is
actually participating in the wait queue of the same lock as the one we
are attempting to acquire.

For this purpose, first introduce accessors to additionally check and
decode tail for the arena slow path. Then, we introduce a new accessor
link_into_waitqueue which abstracts the linking operation. For arena
slow path, this will now be a try_cmpxchg operation. The default value
of node->next will change to the address of the lock whose wait queue
the node is participating in. Using the next field to store the lock
address and synchronize when linking into the wait queue eliminates
races and resolves this ambiguity. The node is reset on the way out
in the arena slow path, so it is abstracted as release_qnode.

Let's consider a scenario where node is being reused (i.e. leaving a
given slow path and entering another slow path). Reuse can from arena to
kernel lock, or kernel to arena lock, or arena to arena, or kernel to
kernel.

First of all, a given CPU will randomly link to a node only when
attempting to acquire an arena lock, as they are susceptible to
corruption. This means that the default value it should expect for
prev->next should be the arena lock it is attempting to acquire.

If the node is moving from kernel to kernel locks, then the cmpxchg will
fail. If the node is moving from arena to kernel lock, we have the
following window.

CPU 0			CPU 1
n0->next = arena
...			cmpxchg(n0->next, arena, n1)
n0->next = NULL

n0->next = kern
...

In this case, next will be overwritten by the NULL assignment and reset,
while the CPU 1 waiter will keep waiting for its MCS locked bit.
Similar outcomes will occur when moving from arena1 to arena2. Moving
from kernel to kernel or kernel to arena is bound to cause failure.

One interesting race window is when CPU 0 forms the queue, and is
waiting for its next pointer to be populated. A valid waiter may be
racing with CPU 1 (which sees CPU 0 due to corruption as prev) to
set the next pointer. However, since both are attempting arena lock
acquisition, both will synchronize over n0->next (n0 == node of CPU 0).
If the corruption-induced cmpxchg accidently wins, it will end up
possibly acquiring the lock. At the same time, the tail will possibly
point to the thread that lost the race (and was the valid waiter).
Incoming threads will then read this value, and attempt to link
themselves into the next pointer of this waiter. They will either
succeed (if they cmpxchg before release_arena_qnode), which will put
them in a spin loop until they timeout, or fail. Eventually,
release_arena_qnode will overwrite their assignment causing any future
cmpxchg to begin failing.

It is possible that between the cmpxchg and reset to NULL, a node's
pointer is visible in the next field. This is harmless, as the compared
value is that of a lock which will never match the address of a node.
The same case occurs when a waiter is already linked and a
corruption-induced cmpxchg happens for a given node, i.e. it ends up
failing.

Thus, in summary, we use the next pointer assignment of the previous
node as the linearization point of all valid and corruption-induced wait
queue linkages. By resetting next to NULL (and using NULL as the default
value for non-arena locks), we ensure corruption-induced linkage
attempts fail, and ensure isolation between arena and non-arena locks.

Whenever arena owners leave the slow path, and observe a non-default
pointer in our node's next pointer, we signal the waiter to stop
spinning and exit the MCS's locked bit waiting loop.

We also introduce a new lockevent (rqspinlock_lock_corrupt) to identify
instances where the lock acquisition fails due to lock word corruption.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>