rqspinlock: Introduce arena rqspinlock kfuncs
Also introduce a bpf_arena_res_spin_lock which wraps over the arena slow
path variant generated separately. Note that this kfunc is not listed in
the kfunc IDs, so it won't be registered or be callable directly.
Instead, the verifier will patch bpf_res_spin_lock calls to this new
bpf_arena_res_spin_lock kfunc based on whether the input to the function
is a PTR_TO_ARENA.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>