bpf: Introduce verifier support for arena rqspinlock
Add support for calling bpf_res_spin_lock kfunc with PTR_TO_ARENA
arguments. The verifier will patch the kfunc call to the correct
underlying implementation (bpf_arena_res_spin_lock). For this purpose,
introduce a get_shadow_kfunc_id helper which informs the verifier
whether there is a potential underlying kfunc that a given kfunc may be
replaced with. In this case, we have to additionally build metadata
using add_kfunc_call and then handle the possible patching in
fixup_kfunc_call.

For managing the lock entries, we don't place constraints on passing the
same arena pointer that was used to acquire a lock, however we do
constrain the user to check the conditional lock result and eventually
call bpf_res_spin_unlock. This is because preemption is disabled by the
lock function, hence the same number of unlock calls must be paired with
successful lock acquisitions.

To avoid requiring the same pointer during unlock, the lock entries
store the id as 0 (default for PTR_TO_ARENA), and the lock ptr as
bpf_arena pointer for the program, which remains the same for all arena
pointers. Thus, on unlock all arena entries will end up matching.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>