x86: Introduce nofault atomic operations
When we access arena pages in the arena locking slow path, the page is
not guaranteed to be populated, as the arena region is sparse and pages
are populated on demand. Regardless, the program can construct an arena
pointer to any arbitrary byte in the heap region, and the page may or
may not be mapped.

To handle such cases, we introduce generic support for atomic operations
that allow handling page faults. We limit exposure to the necessary
atomic operations first (i.e. those listed in atomic_nofault.h), however
support can be extended as necessary in the future.

To handle such page faults, the exception table entry is annotated with
a new type EX_TYPE_KACCESS, which more accurately represents the nature
of addresses accessed by these atomic operations.

The default implementations of these atomic operations in asm-generic
headers simply jump to the fault handling label passed into the atomic
operation. When an architecture is supported, a fault handling atomic
operation will transparently replace this non-functional fallback. Until
then, code using these atomics will compile and function on unsupported
architectures. For now, we add support for atomics used in the
rqspinlock slow path.

While at it, we update the fixup_exception handler to perform default
exception handling for EX_TYPE_KACCESS.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>