rqspinlock: Print a splat on timeout or deadlock
Whenever a timeout and a deadlock occurs, print a message to the dmesg
console, including the CPU where the event occurred, the list of locks
in the held locks table, and the stack trace of the caller, which allows
determining where exactly in the slow path the waiter timed out or
detected a deadlock.

Splats are limited to atmost one per-CPU during machine uptime, and a
lock is acquired to ensure that no interleaving occurs when a concurrent
set of CPUs conflict and enter a deadlock situation and start printing
data.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>